fix logging make repository construction and synchronization best effort
